Adds missing jsdocs, moves types to indicate part of api or internal
