+++ packages/odata-service-inquirer/src/prompts/datasources/cap-project/types.ts
@@ -1,0 +1,31 @@
+import type { CapCustomPaths } from '@sap-ux/project-access';
+import type { CapService, CapServiceChoice, promptNames } from '../../../types';
+import type { ListChoiceOptions } from 'inquirer';
+
+/**
+ * Enumeration of internal prompt names used internally and not supported for modification using OdataServiceInquirerPromptOptions
+ */
+export enum capInternalPromptNames {
+    capProjectPath = 'capProjectPath',
+    capCliStateSetter = 'capCliStateSetter'
+}
+
+/**
+ * Answers to CAP service prompt
+ */
+export interface CapServiceAnswers extends CapService {
+    [promptNames.capProject]: CapProjectChoice['value'];
+    [capInternalPromptNames.capProjectPath]: string;
+    [promptNames.capService]: CapServiceChoice['value'];
+}
+
+export type CapProjectRootPath = {
+    folderName: string;
+    path: string;
+};
+
+export type CapProjectPaths = CapProjectRootPath & CapCustomPaths;
+
+export interface CapProjectChoice extends ListChoiceOptions {
+    value: CapProjectPaths | string;
+}
